+++ Components/Nvim/lua/user/options.lua
@@ -1,0 +1,18 @@
+local option = vim.opt
+-- : help options
+option.cmdheight = 2
+option.conceallevel = 0
+option.fileencoding = "utf-8"
+option.showtabline = 2
+option.smartindent = true
+option.splitbelow = true
+option.splitright = true
+option.shiftwidth = 2
+option.tabstop = 2
+option.cursorline = true
+option.number = true
+option.relativenumber = true
+option.numberwidth = 4
+option.termguicolors = true
+option.guicursor = "n-v-i-c:block-cursor"
+option.shell = "pwsh.exe"